Replaced generating base10-exact floats with approximate comparison.
Finally gave up generating floats that are exactly representable in
base 10 with only a fixed number of digits. It was a fun experiment,
but I don't think it's possible (or at least not too easy) to do
within the entire range of floats. And we do want the tests to cover a
lot of possibilities.

Replaced with approximate matching of floats (recursively).