Support specialized message transcoders
Specialized message transcoders are when the code for
packing/unpacking the message is entirely within the generated code,
As a result, the generated code no longer depends heavily on the gpb
module, but instead contains all the (tailor-made) code needed to
encode/decode the particular messages it knows about.

Additionally the packed option has now gotten its own place in the
`#field{}' record. It is no longer an option in the `#field.opts'
proplist.