Handle msg translations also applicable on top-level
In gpb 4.2.0, the translate_type option is can translate any type of
messages, not just google.protobuf.Any. Hence, translate_any option is
nowadays only a special case of the more general
{translate_type,{{msg,'google.protobuf.Any'},...}.

In gpb 4.3.0, the translate_type option can now also work on the
top-level message to be encoded or decoded, not only for fields or
sub-fields of an existing message.  The gpb_eqc is now adapted to that
fact.